Create DesktopMenu component and the header render it for screen sizes wider than 650px
