+++ src/components/Header/index.module.scss
@@ -6,16 +6,3 @@
   padding: 1.45rem 1.0875rem;
 }
 
-.rightMenu {
-  float: right;
-}
-
-.menuLink {
-  display: inline-block;
-  padding: .25rem;
-}
-
-.menuLink a {
-  color: $off-white;
-  text-decoration: none;
-}
